% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edge_Scenarios.R
\name{addTipChars}
\alias{addTipChars}
\title{Title Add chars to plot}
\usage{
addTipChars(x, r = 1, x.space = 5, legend = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title Add chars to plot
}
\examples{
# patternList()
# EE <- scen$edge.events[[3]]
# edge.color <-edgePatRecode(EE, from='codes', to='cols')
#
# dat <- scen$data[,3:4]
# rownames(dat) <- scen$data[,1]
#
# plot.phylo(scen$tree, edge.color = edge.color, edge.width = 3, show.tip.label = T, no.margin = F, label.offset=10 )
# addTipChars(x=dat, colors, r=1,  x.space=3, legend=TRUE)
# edgelabels(EE, cex = 0.7, frame = "ci", bg = edge.color)
}
