% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_iter_Sinba.R
\name{birth_Des}
\alias{birth_Des}
\title{Title Find minimal edge pattern node for birth-death}
\usage{
birth_Des(EdgeScenarios)
}
\arguments{
\item{EdgeScenarios}{}
}
\value{

}
\description{
Title Find minimal edge pattern node for birth-death
}
