% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\docType{data}
\name{data.sim}
\alias{data.sim}
\title{Tree and two traits for running Sinba}
\format{An object of class \code{list} of length 3.}
\usage{
data(data.sim)
}
\description{
A list with two characters and a tree
}
\keyword{datasets}
