% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edge_Scenarios.R
\name{edgePatRecode}
\alias{edgePatRecode}
\title{Title Recode edge props given patternList()}
\usage{
edgePatRecode(vec, from = "codes", to = "ids")
}
\arguments{
\item{to}{}
}
\value{

}
\description{
Title Recode edge props given patternList()
}
\examples{
# edgePatRecode(vec, from='codes', to='ids')
}
