% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpMat_C}
\alias{ExpMat_C}
\title{Matrix Exponential (scalar implementation)}
\usage{
ExpMat_C(t, A)
}
\arguments{
\item{t}{time}

\item{A}{rate matrix}
}
\value{
probability matrix
}
\description{
Matrix Exponential (scalar implementation)
}
