% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edge_Scenarios.R
\name{getEdgeScenarios}
\alias{getEdgeScenarios}
\title{Title Create edge scenarios}
\usage{
getEdgeScenarios(chars, focal.states = c(1, 1), tree, token.maps)
}
\arguments{
\item{token.maps}{}
}
\value{

}
\description{
Title Create edge scenarios
}
