% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_iter_Sinba.R
\name{getPhyloCTMC_distr}
\alias{getPhyloCTMC_distr}
\title{Title Calculate analtical distribution of site patterns at two-branch tree}
\usage{
getPhyloCTMC_distr(Qsim, t)
}
\arguments{
\item{Qsim}{rate matrix}

\item{t}{vector of times for each of 2 branches}
}
\value{
list of site patterns in binary format and probability for each pattern
}
\description{
Title Calculate analtical distribution of site patterns at two-branch tree
}
