% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_iter_Sinba.R
\name{infer2branchTrees}
\alias{infer2branchTrees}
\title{Title Run batch inference on two-branch tree}
\usage{
infer2branchTrees(sim.multi, Qsim, Qcom, t, pi, add.TIC = FALSE)
}
\arguments{
\item{pi}{}
}
\value{

}
\description{
Title Run batch inference on two-branch tree
}
