% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRateMatrices.R
\name{iter_new_Qcom}
\alias{iter_new_Qcom}
\title{Title Create new Qcom by merging paramter estimates that are similar, and setting 0 estmates to 0}
\usage{
iter_new_Qcom(pars, Qcom.cor)
}
\arguments{
\item{Qcom.cor}{Qcom}
}
\value{

}
\description{
Title Create new Qcom by merging paramter estimates that are similar, and setting 0 estmates to 0
}
