% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Sinba.R
\name{likelihood_Sinba}
\alias{likelihood_Sinba}
\title{Title Calculate likilehood for Sinba Model (coorelated evolution of two characters)}
\usage{
likelihood_Sinba(
  p,
  phy,
  liks,
  rate,
  root.p,
  Qindv.zero,
  Pindv,
  edge.pattern.focal,
  cub.method = "pcubature"
)
}
\arguments{
\item{cub.method}{}
}
\value{

}
\description{
Title Calculate likilehood for Sinba Model (coorelated evolution of two characters)
}
