% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRateMatrices.R
\name{mk_Pindv}
\alias{mk_Pindv}
\title{Title}
\usage{
mk_Pindv(
  Qcom,
  p1 = c(0.5, 0, 0.5, 0),
  p2 = c(0.5, 0.5, 0, 0),
  p0 = c(1, 0, 0, 0)
)
}
\arguments{
\item{names}{}
}
\value{

}
\description{
Title
}
