% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRateMatrices.R
\name{mk_QcomPartitions8}
\alias{mk_QcomPartitions8}
\title{Title Get all parametrizations of Q with 8 elements}
\usage{
mk_QcomPartitions8()
}
\value{

}
\description{
Title Get all parametrizations of Q with 8 elements
}
