% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferSinba.R
\name{runSinba}
\alias{runSinba}
\title{Title Infer corrleated evolution of two characters using Sinba model}
\usage{
runSinba(
  phy,
  data,
  edge.pattern = NULL,
  Edge.Scenarios,
  root.p = NULL,
  Qindv,
  Pindv = NULL,
  p = NULL,
  ip = NULL,
  lb = 0,
  ub = 100,
  cub.method = "pcubature",
  diagn = FALSE,
  optim.method = "subplex",
  subplex.rounds = 1,
  subplex.n.cores = 1L,
  gensa.its = 100
)
}
\arguments{
\item{do.thorough}{}
}
\value{

}
\description{
Title Infer corrleated evolution of two characters using Sinba model
}
\examples{

}
