% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Phy.R
\name{simDarwin}
\alias{simDarwin}
\title{Title Simulate Darwin Scenario}
\usage{
simDarwin(
  ntips = 100,
  n.trees.sim = 10,
  Upper.otu.bound = 0.6,
  Lower.otu.bound = 0.4
)
}
\arguments{
\item{Lower.otu.bound}{}
}
\value{

}
\description{
Title Simulate Darwin Scenario
}
