% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_iter_Sinba.R
\name{simulate_ExpLinLn}
\alias{simulate_ExpLinLn}
\title{Title Simulate a two-branch tree with characters}
\usage{
simulate_ExpLinLn(Q, t = c(1, 1), n.samples = 100, pi = c(1/4, 1/4, 1/4, 1/4))
}
\arguments{
\item{Q}{rate matrix}

\item{t}{vector of times for each of 2 branches}

\item{n.samples}{N of characters}

\item{pi}{intial vecotr}
}
\value{
list of two chars in binary mode (y- N states, x- N chars)
}
\description{
Title Simulate a two-branch tree with characters
}
