% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Phy.R
\name{simUnrepBurst}
\alias{simUnrepBurst}
\title{Title Simulate Unreolicated Burst}
\usage{
simUnrepBurst(
  ntips = 100,
  n.trees.sim = 10,
  Upper.otu.bound = 0.6,
  Lower.otu.bound = 0.4,
  Q,
  p.root = c(1, 0, 0, 0),
  freq.threshhold = 5
)
}
\arguments{
\item{freq.threshhold}{}
}
\value{



}
\description{
Title Simulate Unreolicated Burst
}
\examples{
Title Simulate Unreplicated Burst

}
